Fix Rotate

+++ PdfFile/SrcWriter/Annotation.cpp
@@ -1346,7 +1346,7 @@ namespace PdfWriter
 	CStampAnnotation::CStampAnnotation(CXref* pXref) : CMarkupAnnotation(pXref, AnnotStamp), m_pAPStream(NULL)
 	{
 	}
-	void CStampAnnotation::SetRotate(int nRotate)
+	void CStampAnnotation::SetRotate(double nRotate)
 	{
 		Add("Rotate", nRotate);
